requirements.txt
Flask==3.1.2
pytesseract==0.3.13
Pillow==11.3.0
pyttsx3==2.99
gunicorn
